Upgrade lombok 1.18.24 -> 1.18.34 (java 21 support) (#68)
We choose lombok 1.18.34 like in springboot-dependencies 3.3.3
that we are currently using

Maybe we should delegate entirely to springboot the version of
lombok? But contrary to other dependencies, lombok is a build only
tool tied to our source code so it makes sense to handle differently
but in practice we didn't upgrade it until we were forced to by java 21

Not sure why it was added initialy by me in
https://github.com/gridsuite/dependencies/commit/9abee85fe4585a6bca2dc0f18261045f09e9ea0b
maybe just because it's a compile tool.

Signed-off-by: HARPER Jon <jon.harper87@gmail.com>